// TS/src/types/other.ts
export interface BlogType {
  id: string | number
  _id: string; // <-- not `id`
  title: string;
  description: string;
  image: string[];
  name: string;
  createdAt: string;
  category: {
    name: string;
    variant: string;
  }
  projectFile?: string
}